SPMS-235 added premium tests
