+++ java8-okhttp-gson/src/test/java/com/sphereon/sdk/pdf/api/advanced/AdvDocxXlsxConversion2PDFApiTest.java
@@ -1,7 +1,9 @@
 package com.sphereon.sdk.pdf.api.advanced;
 
 import com.sphereon.sdk.pdf.api.Conversion2PDFApiTest;
+import com.sphereon.sdk.pdf.api.util.DocxXlsxSetupAndPdfAssert;
 import com.sphereon.sdk.pdf.api.util.PDFReadUtils;
+import com.sphereon.sdk.pdf.api.util.SetupAndPdfAssert;
 import com.sphereon.sdk.pdf.handler.ApiException;
 import com.sphereon.sdk.pdf.model.ConversionSettings;
 import org.apache.commons.lang3.StringUtils;
@@ -19,11 +21,11 @@ import java.io.IOException;
 public class AdvDocxXlsxConversion2PDFApiTest extends Conversion2PDFApiTest {
 
     private final boolean WRITE_TO_FILE = true;
+    private SetupAndPdfAssert setupAndPdfAssert = new DocxXlsxSetupAndPdfAssert();
 
     @Before
     public void setUp() {
-        super.uploadConfigs.add(new FileUploadConfig("11Sheets.xlsx",Conversion2PDFApiTest.class.getResource("/11Sheets.xlsx")));
-        super.uploadConfigs.add(new FileUploadConfig(2, "4Pages.docx", Conversion2PDFApiTest.class.getResource("/4Pages.docx")));
+        super.uploadConfigs.addAll(setupAndPdfAssert.getUploadConfigs());
         super.engine = ConversionSettings.EngineEnum.ADVANCED;
         super.writeToFile = WRITE_TO_FILE;
     }
@@ -60,45 +62,7 @@ public class AdvDocxXlsxConversion2PDFApiTest extends Conversion2PDFApiTest {
 
     @Test
     public void _70_checkGeneratedPdf() throws IOException{
-        if (WRITE_TO_FILE) {
-            PDDocument doc = null;
-            try {
-                doc = PDDocument.load(new File(Conversion2PDFApiTest.PDF_DOWNLOAD_PATH));
-                String allText = PDFReadUtils.getTextFromPdfPage(doc,1,19);
-                Assert.assertEquals(19, doc.getNumberOfPages());
-                Assert.assertTrue(PDFReadUtils.getTextFromPdfPage(doc,1,1).contains("Estimating template guidelines"));
-                Assert.assertEquals(1,StringUtils.countMatches(allText,"Estimating template guidelines"));
-                Assert.assertEquals(2,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,1,19),"Pre-estimate checklist"));
-                Assert.assertEquals(2,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,2,2),"Pre-estimate checklist"));
-                Assert.assertEquals(2,StringUtils.countMatches(allText,"Project Assumptions"));
-                Assert.assertEquals(2,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,3,3),"Project Assumptions"));
-                Assert.assertEquals(1,StringUtils.countMatches(allText,"Project Estimate Summary"));
-                Assert.assertEquals(1,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,4,4),"Project Estimate Summary"));
-                Assert.assertEquals(1,StringUtils.countMatches(allText,"Effort estimate"));
-                Assert.assertEquals(1,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,5,5),"Effort estimate"));
-                Assert.assertEquals(1,StringUtils.countMatches(allText,"Resource type estimate checklist"));
-                Assert.assertEquals(1,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,5,5),"Resource type estimate checklist"));
-                Assert.assertEquals(1,StringUtils.countMatches(allText,"Indirect estimate checklist"));
-                Assert.assertEquals(1,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,5,5),"Indirect estimate checklist"));
-                Assert.assertEquals(1,StringUtils.countMatches(allText,"Hardware Cost estimate"));
-                Assert.assertEquals(1,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,6,6),"Hardware Cost estimate"));
-                Assert.assertEquals(1,StringUtils.countMatches(allText,"Software Cost estimate"));
-                Assert.assertEquals(1,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,7,7),"Software Cost estimate"));
-                Assert.assertEquals(1,StringUtils.countMatches(allText,"Network Cost estimate"));
-                Assert.assertEquals(1,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,8,8),"Network Cost estimate"));
-                Assert.assertEquals(1,StringUtils.countMatches(allText,"On-going support estimate"));
-                Assert.assertEquals(1,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,9,9),"On-going support estimate"));
-                Assert.assertEquals(3,StringUtils.countMatches(allText,"Admin costs not in overhead"));
-                Assert.assertEquals(2,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,10,10),"Admin costs not in overhead"));
-                Assert.assertEquals(2,StringUtils.countMatches(allText,"Risk & Change"));
-                Assert.assertEquals(1,StringUtils.countMatches(PDFReadUtils.getTextFromPdfPage(doc,11,11),"Risk & Change"));
-                Assert.assertEquals(96, StringUtils.countMatches(allText,"Lorem ipsum dolor sit amet"));
-                Assert.assertEquals(2, PDFReadUtils.getImagesFromPdfPage(doc, 13).size());
-                Assert.assertEquals(2, PDFReadUtils.getImagesFromPdfPage(doc, 17).size());
-
-            } finally {
-                doc.close();
-            }
-        }
+        if (WRITE_TO_FILE)
+            setupAndPdfAssert.checkPdfFile(Conversion2PDFApiTest.PDF_DOWNLOAD_PATH);
     }
 }
