+++ java8-okhttp-gson/src/test/java/com/sphereon/sdk/pdf/api/premium/PrePngXlsDocConversion2PDFApiTest.java
@@ -1,32 +1,28 @@
-package com.sphereon.sdk.pdf.api.advanced;
+package com.sphereon.sdk.pdf.api.premium;
 
 import com.sphereon.sdk.pdf.api.Conversion2PDFApiTest;
-import com.sphereon.sdk.pdf.api.util.PDFReadUtils;
-import com.sphereon.sdk.pdf.api.util.PngPdfTxtPptSetupAndPdfAssert;
+import com.sphereon.sdk.pdf.api.util.PngXlsDocSetupAndPdfAssert;
 import com.sphereon.sdk.pdf.api.util.SetupAndPdfAssert;
 import com.sphereon.sdk.pdf.handler.ApiException;
 import com.sphereon.sdk.pdf.model.ConversionSettings;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.pdfbox.pdmodel.PDDocument;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
 
-import java.io.File;
 import java.io.IOException;
 
 
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class AdvPngPdfTxtPptConversion2PDFApiTest extends Conversion2PDFApiTest {
+public class PrePngXlsDocConversion2PDFApiTest extends Conversion2PDFApiTest {
 
     private final boolean WRITE_TO_FILE = true;
-    private SetupAndPdfAssert setupAndPdfAssert = new PngPdfTxtPptSetupAndPdfAssert();
+    private SetupAndPdfAssert setUpAndPdfAssert = new PngXlsDocSetupAndPdfAssert();
+
     @Before
     public void setUp() {
-        super.uploadConfigs.addAll(setupAndPdfAssert.getUploadConfigs());
-        super.engine = ConversionSettings.EngineEnum.ADVANCED;
+        super.uploadConfigs.addAll(setUpAndPdfAssert.getUploadConfigs());
+        super.engine = ConversionSettings.EngineEnum.PREMIUM;
         super.writeToFile = WRITE_TO_FILE;
 
     }
@@ -57,6 +53,6 @@ public class AdvPngPdfTxtPptConversion2PDFApiTest extends Conversion2PDFApiTest 
     @Test
     public void _70_checkGeneratedPdf() throws IOException{
         if (WRITE_TO_FILE)
-            setupAndPdfAssert.checkPdfFile(Conversion2PDFApiTest.PDF_DOWNLOAD_PATH);
+            setUpAndPdfAssert.checkPdfFile(Conversion2PDFApiTest.PDF_DOWNLOAD_PATH);
     }
 }
